Move specific sockets under services
