+++ src/assets/src/hooks/useWebSocket.ts
@@ -1,7 +1,5 @@
 import { useState, useEffect } from "react";
 
-import { QueueHost, QueueAttendee, User, MyUser } from "../models";
-
 interface OfficeHoursMessage<T> {
     type: "init"|"update"|"deleted";
     content: T;
@@ -40,26 +38,4 @@ export const useWebSocket = <T>(url: string, update: (content: T) => void, handl
         }
     }, []);
     return error;
-}
-
-export const useQueueWebSocket = (queue_id: number, update: (q: QueueHost | QueueAttendee | undefined) => void) => {
-    return useWebSocket(
-        `ws://${location.host}/ws/queues/${queue_id}/`,
-        update,
-    );
-}
-
-export const useUsersWebSocket = (update: (users: User[]) => void) => {
-    return useWebSocket(
-        `ws://${location.host}/ws/users/`,
-        update,
-        false,
-    );
-}
-
-export const useUserWebSocket = (user_id: number, update: (user: User|MyUser) => void) => {
-    return useWebSocket(
-        `ws://${location.host}/ws/users/${user_id}/`,
-        update,
-    );
 }
